feat: change public interface lift the key exchange outside the proxy
In many reverse proxy applications the target host depends on the
metadata or identity of the incoming request. To suite this case,
the calling code needs access to the underlying *ssh.ServerConn
which contains relevant metadata including the `user`. To serve this
need, this commit adjusts the signiture of `(*ReverseProxy).Serve`.